versatile-data-kit: Remove references to Data Lake/Data Warehouse
The VDK repo contained many references to the concepts of a Data
Lake or a Data Warehouse; however these terms were not defined
anywhere and were not represented in code. For the purposes of
clarity, this change removes all references to those two concepts
and replaces them with references to a generalized concept of a
database.

Signed-off-by: gageorgiev <gageorgiev@vmware.com>