vdk-heartbeat: add flag to disable manual execution test
We should make vdk-heartbeat very configruable since we do not want to
be able to sue in very different deployments and setups of Versatile
Data Kit. It is possible that in some setups manual execution is not
important and testing for it is not necessary or it is to  expensive (if
the job under test is very slow). I also would use it to skip this check
unitl we fix a bug in execution api with too long names in short term.

Testing Done: ran vdk-heartbeat with export
check_manual_job_execution=True and verified in logs that manual
execution was skipped

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>