vdk-sqlite: Handle column mismatch error
When using SQLite databases, the database table needs
to be created before data is inserted into it. In cases
when there is a mismatch between column names in the input
payload and the column names of the database table, the
vdk-sqlite plugin did not handle properly the exception
produced by the SQLite connector.

This change extends the error handling of the plugin, so
that exceptions caused when there is a difference in the
number of columns, or a mismatch between column names, is
reported as a User Error, and not a Platform Error.

Testing Done: Added unit test

Signed-off-by: Andon Andonov <andonova@vmware.com>