Fix failed unit test
Test was failing because we changed the logic behind updating
data job executions. The test was trying to change a finished
execution to running state. We discussed this and agreed we
should not be allowing this behaviour.

Changed the test to update a different execution in running
state. Added another test which initialises another execution
to a running state, and modified the failing test to update
it to a Skipped status.

Signed-off-by: Momchil Zhivkov <mzhivkov@vmware.com>