vdk-control-cli: provide option to disable authentication
Extend the set-default command to provide an option (--authentication-disable)
to disable authentication. If either this new option is set or the
VDK_AUTHENTICATION_DISABLE environment variable is set to True,
the authentication will be disabled.

Testing done: Manually set the new --authentication-disable option
and verified that authentication is indeed disabled in subsequent
vdk commands that use authentication.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>