vdk-server: disable authentication during server installation
We don't want to require users to authenticate when installing
local Control Service (in which case authentication is disabled
on the server side anyway). This commit turns off authentication
in the vdk client during local Control Service installation so
that the cli is useable immediatelly without the necessity for
further configurations.

Testing done: Installed the Control Service locally and
verified that the vdk is useable against it without
authentication problems.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>