control-service: update the prometheus rules
This commit fixes some problems with the Prometheus rules that were
discovered and addressed in one of the Control Service adoptions.

In addition, the time that a JobDelay alert is active before it
becomes firing is increased to 5 minutes to avoid some false
positives when the schedule_cron of an existing job is changed.

Testing done: Installed the Helm chart in a staging environment
and verified that the Prometheus rules installed correctly. The
alerts themselved produced by these rules are test as part of one
of the Control Service adoptions.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>