control-service: implement Execution Logging API
We need Logging API for current executions - especially in the absence
of log aggregators (like fluent/kibana). See more details on reasoning
in PR 259.

The current implementation would just get from the last pod of the job
execution (hence last attempt) logs. And if tail_lines is set it will
get last N lines.

Testing Done: unit tests and integration test (for which I resisted the
template to refactor it but we need to simplify and split it.)

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>