control-service: Convert debug logs to info
When a requirements.txt file of a data job is broken,
a platform error is raised, because the Control Service
is not able to get the logs from the builder k8s pod.

Currently, we cannot properly investigate what are the
reasons for that, as we don't log enough data in the control
service's log stream.

This change converts some of the debug log messages to info messages
to allow for a better view over the possible reason for the service
not being able to get the logs from the builder pods.

Testing Done: Logging change. New logs appear in the Control Service
log stream.

Signed-off-by: Andon Andonov <andonova@vmware.com>