control-service: sync job execution statuses
We need to keep in sync all job executions in the database
in case of Control Service downtime or missed Kubernetes Job Event.

This change aims to synchronize Data Job Execution statuses in the database
with the actual running jobs in Kubernetes.

It takes into account the executions that are started concurrently
during the synchronization or delayed Kubernetes Job Events
by synchronizing only executions that are started before now() - 3 minutes.

Testing done: unit tests

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com