control-service: add executions logging api (#259)
* control-service: add executions logging api

If a user installs a control serivce/versatile data kit by default they
do not have access to execution logs. This is likely to be huge issue
for adoptoin. Requiring that users integrate with tools like fluentd for
logging or having them expose kubectl can be pretty hard.

All APIs that support some kind of
task execution naturally need to expose logs interface to users.
Examples:
 - Lambda (using aws cloudwatch)
 - Data Bricks Jobs (API returns link to spark ui/logs)
 - airflow (rest api)
 - OpenFaas (has faas-cli logs)

Initially it would be used
- in vdk-server workflow - instead of having users use kubectl, let
them use vdk execute --logs and we can remove any mention of
kind/kubectl.
- in vdk-heartbeat - root cause of vdk-heartbeat test can be very
complex. After the test finishes it cleans after itself. And this causes
all logs of that execution to disappear making it very hard to see why
the data job started by the test has failed.

Currently this is added as only experimental feature. There's some
issues that need to be decided before it can be promoted to stable (will
be noted as TODOs in the code)

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>

* Update projects/control-service/projects/pipelines_control_service/src/main/java/com/vmware/taurus/datajobs/DataJobsExecutionController.java