vdk-core: Remove databases from DB_DEFAULT_TYPE description
The DB_DEFAULT_TYPE env variable description previously had a
section which would in the future list the database connectors
available through plugins, which was currently labeled TODO.
However, since the ManagedConnectionRouter, which is where the
list of connectors is stored, is part of the JobContext, and
the DB_DEFAULT_TYPE description is set through a vdk_configure
call, there is no simple way of capturing and display the list.
For this reason, the change removes the section from the
description for the mean time, but keeps the TODO code comment.

Testing done: pipelines passed

Signed-off-by: gageorgiev <gageorgiev@vmware.com>