vdk-control-cli: add vdk execute --logs
After engineers have started manually their data job it will be good for
them to be able to see their logs in order to track the progress of the
job and if the job fails to troubleshoot and so on. In other words Logs
are invaluable.

Testing Done: unit tests added.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>