vdk-server: attempt to address a problem on some enviroments
On some environments installation of the ingress component
results in the following error:
> failed calling webhook "validate.nginx.ingress.kubernetes.io": Post
> https://ingress-nginx-controller-admission.ingress-nginx.svc:443/networking/v1/ingresses?timeout=10s:
> dial tcp 10.96.170.241:443: connect: connection refused

While the origin of this error cannot be discovered, suggested
workaround is to delete the validatingwebhookconfigurations
object.

As part of this commit, we remove this object altogether.

Testing done: Performed the `vdk server -i` installation and verified
that it works without the removed configuration.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>