vdk-core: Raise exception if Data Job has no steps
Currently, an empty Data Job, i.e. one with no steps, will succeed.
This is suboptimal as in certain cases a user might believe they
are executing a Data Job while in fact they are not.
There is no apparent use case for an empty Data Job.

This change fixes the aformentioned problem by making VDK raise an
exception if there are no steps in the target Data Job.

Testing done: executed an empty Data Job and verified an exception
is raised

Signed-off-by: gageorgiev <gageorgiev@vmware.com>