control-service: Add vdk-heartbeat image
As part of the the CI/CD process, we could benefit from
having the vdk-heartbeat package built as a docker image.

This change adds a dockerfile to facilitate building the
vdk-heartbeat docker image when a new release of the
Control Service is made.

Testing Done: Built the image in docker locally and
ran it to verify that it is being executed as expected.

Signed-off-by: Andon Andonov <andonova@vmware.com>