vdk-heartbeat: make sure trino job install vdk-trino (#284)
The default vdk we use quickstart-vdk comes only with sqlite driver. For
trino data job test we need to install teh vdk-trino plugin and we do it
in teh requirements.txt of the job. This has the added benefit of
testing requirements.txt functionality is working correctly which is
pretty nice.

Testing Done: vdk-heartbeat -f trino_based_heartbeat_config.ini

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>