vdk-logging-json: Format logs before initializing job (#286)
Before, the change of the log format happens in a initialize_job
hook. This is suboptimal, as some logs are printed before a job
is initialized.
To address this, this change includes another hook invocation,
namely vdk_start, to format logs as early as possible. The early
logs have their jobname and attemptid fields empty, as those are
part of the JobContext, which is unreachable at that point of the
execution.

Testing done: verified logs in local execution, pipelines passed

Signed-off-by: gageorgiev gageorgiev@vmware.com