control-service: give Control Service permission to read logs
In order to be able use the Logging API (execute/logs) Control Service
need to be able to read them from the pod. When installing the helm
chart we need to make sure that the control service has permission to
read the pod logs. Until now it did not and so no logs were returned.

The same goes for builder jobs. Recently we have had issues with Control
Service not being able to read builder logs which I suspect is a
permission issue.

Testing Done: vdk execute --logs -n ai-job -t ai - returned logs after
setting the permissions. Before that it was not.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>