vdk-heartbeat: show last job execution logs
When vdk heartbeat fails , often it fails because the data job execution
started during the test has some issues. But it is very hard to
troubleshoot since developers would need to go to the kubernetes cluster
and check out the pod logs manually. Also it needs to be done while the
test is runniing as they'd expire/disappear after the test is finished
in noraml circumstances since the test clean up after itself.

We are adding a step in hearbeat to print the logs of the last job
execution - just after run test has finished and also in case of a
failure before job is deleted.

Testing Done: vdk-heartbeat -f local.ini and saw the logs.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>