vdk-control-cli: deploy -p $dir should not fail if team is not set
vdk deploy -p $job_dir has a logic to read team name from the config.ini
file and should not generally fail.
Also even if team is not set in config.ini it should prompt for team
name and gain it should not fail

I fixed that by removing it asrequired and explicitly prompting for team
when necessary. This way Deploy create would prompt when necessary.

Testing Done: ran vdk deploy -p . and it prompted me for team instead of
failing.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>