vdk-server: remove pod limits
I am setting all pod limits to 0 which effectively removes them.
Allowing to start more containers. This is still limitted by the memory
of the node (the user laptop in our case) and kubernetes would evict if
not is under pressure. See discussions in
https://github.com/kubernetes/kubernetes/issues/86244

So I think it would be ok and would allow users on any machine to use
it. Laptop would little memory would quickly reach the node limits
otherwise. I reached them (with 16G) on the first job execution.

Testing Done:
vdk server -i locally and saw  (kubectl get all) everything is created
as it should be and working correctly. Started successfully execution

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>