vdk-server: increase job execution pod retention
With vdk-server, the only way to inspect the fail or successful job is
through kubernetes. The job logs are not saved to aggregated place. So
we need a bit bigger history. As vdk-server would primarily be used for
personal use and or testing then having bigger history would not be an
issue.

Testing Done: vdk-server -i and vdk deploy and verified the cronjob is
created with porper history limits with kubectl describe

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>