control-service: Format logs in JSON
Control Service logs must be formatted in JSON to be able to be
parsed correctly in Kibana. This change adds an encoder which
formats logs in JSON for an instance of the Control Service with
the LOGGING_SYS_LOG_HOST env variable defined.

Testing done: ran CS with the encoder set for local logs

Signed-off-by: gageorgiev <gageorgiev@vmware.com>