control-service: implement authorization (#294)
As part of our authorization logic, we would like to provide means to
authorize users against a predefined set of claim values (e.g. organization
IDs) and roles in addition to the webhook-based one.

These claim values and roles are provided as configurations
at service startup and are subsequently used to authorize the users.
Both claim values and roles are optional and if not provided, the
respective validation is not performed. The new authorization logic
is complementary to the existing webhook authorizations. If they
are both enabled, the claim values and roles are validated first.

Testing done: Integration tests passed. Locally run the service configured
with the new authorization settings and validated that a user of the appropriate
organization and with the appropriate roles is authorized successfully.
Similarly, a user outside of the allowed organizations or without the
required roles is not authorized.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>