vdk-heartbeat: improve logging
It is not very clear why fail test if it has misconfigured test run
module name or class. This makes it a bit more clear.

Also job logs were printed on a single line (new line was escape) and we
un-escape it here

Testing Done: vdk-heartbeat -f local.ini and saw proper logs. try with
misconfigured and correctly configure configuration.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>