control-service: fix bug in bad deployment metadata
Deployment metadata is kept in two places in the job resource which is
not necessary and also error prone.
This reads lastDeployedBy/Date from the new location set by
JobAnnotation

The same goes for git commit hash. It has not been noticed because we
try to read it from the image name.
But if the image name contains registry with port e.g
localhost:5000/job-name:hash-name - the logic breaks and we get job
version "5000/job-name" instead of expected hash name.

Testing Done: extended integration test to explcitly verify correct
metadata

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>