vdk-core: Validate plugin entry point during build stage
Currently, unless explicitly tested, there is no step validating
that the entry point of a plugin is properly set. This change
fixes this by including an extra step in the build stage, which
runs `vdk version`, then greps through the output for the plugin
name, and if it doesn't find it, it exits.

Signed-off-by: gageorgiev <gageorgiev@vmware.com>