control-service: improve list executions api correctness
The VDK Skip plugin relies heavily on running execution status.
In some cases, the execution status in the database may be
outdated due to delay. As a result the next job execution will be skipped.

In order to avoid such cases we must return only running jobs
in Kubernetes when a filter (status=RUNNING) is passed to the API.

Testing: unit tests

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com