plugins: fix entry-point check
The entry point check was failing for vdk-test-utils and quickstart-vdk
which are not actual plugins.
quickstart-vdk is a distribution of vdk (aggregation of plugins) to get
started quickly .
vdk-test-utils is a library for testing plugins or vdk-core.

So they are added as expcetions explicitly to the check.

Also decoupled the building of plugins in separate plugin to make it
easier to develop and reuse locally

Testinng Done: This change CI and locally from quickstart-vdk,
vdk-ingest-file and vdk-test-utils directory run `../build-plugin.sh`
successfully

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>