vdk-plugin-control-cli: introduce job execution skipping functionality (#321)
why: As a user, when starting a data job execution manually through
the new execution API I don't want a scheduled job to run in parallel
with my manual execution. This is why we need a plugin that is able
to check for a different running execution and, skip it if necessary.

what: The new implementation will make a call to the control-service
API and query for any executions in RUNNING or SUBMITTED state.
If we have such executions present, it will SKIP the current execution.
Added the implementation to the vdk-plugin-control-cli existing
plugin as discussed. The entry point for the new functionality is through
the run_job hookimpl as we need log_config_type, job_name, execution_id
and job_team to be initialized by the initialize_job hooks. We also have the
tryfirst param enabled.

testing: Verified skipping logic is working as intended.
Added unit tests. CI/CD

Signed-off-by: Momchil Zhivkov mzhivkov@vmware.com