control-service: extend pod termination message
In order to be able to use the K8S Pod termination message
to transfer various data (such as vdk version) we need to switch
from plain text to json format of termination message.

This change is backward compatible with the latest VDK version
that uses plain text format.

Testing: unit tests

Signed-off-by: Miroslav Ivanov miroslavi@vmware.com