control-service: change CI/CD release new version on each PR merge
Up until now Control Service releases had to be done manually
by making changes to several files and then making a new PR
and merge it.

With this commit a new version of the Control Service is
released automatically every time a new image of the Control Service
is built, which happens every time a PR with changes to the code
base of the Control Service is merged. The MAJOR and MINOR components
of the version are taken from the version.txt file. The PATCH
component is automatically generated.

The CONTRIBUTING.md file was updated to reflect the new release
process. The CHANGELOG.md file was cleaned up.

Testing done: none

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>