vdk-trino: Ensure tests would catch the 'LIKE view' issue
A previous PR fixed an issue with the SCD1 template attempt to
create a table LIKE a view, which isn't permitted. This change
ensures that the SCD1 template test would catch that issue.

Testing done: ensured that the new test succeeds on the fixed
template, but fails on the old template

Signed-off-by: gageorgiev <gageorgiev@vmware.com>