control-service: remove the temporary testing components
Testing done: Changed the steps of interest to execute on PR rather
than on merge and verified that they run correctly, i.e. verified that
the version of the used control service image is correct, that the
control service version itself is properly generated and that the
heartbeat is released with the appropriate version.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>