control-service: change CI/CD to release a new version on each PR merge (#357)
Up until now, Control Service releases had to be done manually
by making changes to several files and then making a new PR
and merge it.

With this commit, a new version of the Control Service is
released automatically every time a new image of the Control Service
is built, which happens every time a PR with changes to the code
base of the Control Service is merged. The MAJOR and MINOR components
of the version are taken from the version.txt file. The PATCH
component is automatically generated.

Updated the CONTRIBUTING.md file to reflect the new release
process. Cleaned up the CHANGELOG.md file.

Testing done: Changed the steps of interest to execute on PR rather
than on merge and verified that they run correctly, i.e. verified that
the version of the used control service image is correct, that the
control service version itself is properly generated and that the
heartbeat is released with the appropriate version.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>