vdk-logging-json: Allow for logging of the stack trace
Until now, formatting a stack trace log was not possible, as it
uses a separate function formatException() instead of format().
This change fixes that by overriding the aformentioned function,
capturing the stack trace, replacing the newlines and double quotes,
formatting it in JSON and logging it.

Testing done: ran job locally and verified the stack trace is JSON
formatted

Signed-off-by: gageorgiev <gageorgiev@vmware.com>