control-service: fix vdk-heartbeat
The post-deployment test(vdk-heartbeat) seems to be failing with
authorization errors. I guess it is because of misconfiguration.

I set the correct VDK_OPTIONS file in Gitlab CI (it can be overridden by
Gitlab CI Variable).
I think it might be better to leave it and edit it directly in Gitlab CI
? And keep the one in source control for reference purposes.

Testing Done: I set VDK_OPTIONS file in
https://gitlab.com/vmware-analytics/versatile-data-kit/-/settings/ci_cd
and ran manually to deploy control service with new options in pipeline
https://gitlab.com/vmware-analytics/versatile-data-kit/-/pipelines/386027936

And then ran post-deployment test in same pipeline

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>