control-service: fix false-positive alert notifications
We are experiencing intermittently email alerts for job executions
that are false positive (i.e. for executions that have happened
in the past or for user errors for executions that have succeeded).

This is the first of several commits that are an attempt to remedy
these.

This particular commit improves logging during data job execution
monitoring and adds methods for gauge removal that are not yet
fully utilized.

Testing done: unit and integration tests pass

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>