vdk-core: configure correct attempt, execution and op id.
For a configuration variable to be used it needs to be defined with
ConfigurationBuilder#add. If it is not defined the environment variables
configuration provider will likely not detect it.

In our case we define attempt id using:

config_builder.add(CommonStoreKeys.ATTEMPT_ID.key, None)
...

This is the configuration store key which as string look like this:

    ATTEMPT_ID: StoreKey[str] = StoreKey[str]("vdk.attempt_id")

So we basically defined it as `vdk.attempt_id` and the (environment
variables) configuration provider appends VDK_ (and replaces dot with an
underscore). So it searches for
the environment variable named "VDK_VDK_ATTEMPT_ID" in this example.

execution_id and op_id already had correct definitions. In this change
we adding one for op id. And removing redundant ones.
Also try to improve their description a bit.

Testing Done: add new test. I decide to use functional type of test
since the logic spans more than one vdk core component/class instead of
a unit test.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>