control-service: improve monitoring code
This is the second in a series of changes that aim to fix
false-positive alerts occasionally sent by the control service.

This particular commit refactors the monitoring code:

The metrics management is now governed by a dedicated class -
DataJobMetrics. Service classes no longer reference the
monitors classes and instead operate on the metrics directly
via this new class. Some synchronizations logic has been
forfeited in order to simplify the code and instead rely on
eventual consistency via the *MonitorSync classes.

The logging has been improved to allow easier troubleshooting
of issues.

In a future commit I plan to remove the DataJobInfoMonitor class
and merge it in the DataJobStatusMonitor. Also, I will extend
the MonitorSync classes to also clean up gauges of removed data jobs.

Testing done: There are new unit tests. Unit and integration
tests passed.
Manually tested the metrics classes by running the service
locally and creating and deploying a data job and observing that
the exposed metrics are correct.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>