vdk-core: vdk config-help usability improvements
It seems a lot of people miss how configuration variables can be set
when using vdk. On top of the help we have described the configuration
providers (environment variables or job conifg.ini file) but they are
usually hidden by the long list of variables. To fix that we print them
on the end as well as in the beginning.

Second - different plugins can provider configuraiton variables and we
have not control over how they'd format them . E.g if someone uses
multiline strng often they'd end up having a lot of long whitespaces .
```
The maximum number of payloads that can be stored
                                        for ingestion at any
given time. When limit
is reach send_*_for_ingestion
method would effctively block
unitl the queue is freed.
```

which is slightly hard to read. We trim those

Third - some variables name can be pretty long. We use the first 40
characters for the var name but if it is longer than 40 it get mixed wit
hteh description. We add explicitly 2 spaces if that happens.

Testing Done: ran locally vdk config-help

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>