control-service: ignore late Kubernetes event (#409)
Occasionally, for no obvious reasons, our watching logic recieves out
of band events for jobs that have completed long time ago. This disrupts
our alerting if these events are for jobs that are different than the
last job, as this triggers an alert.

This commit introduces a check for such events and ignores them. More
specifically, events that are received later than one hour after the job
has completed are ignored.

Testing done: This particular change is hard to test manually due to its
nature. So I rely on the unit and integration tests.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>