vdk-core: fail job if job template fails (#416)
If the template executed using job_input.execute_template fails we do
not raise exception. Instead  job_input.execute_template function
returns ExecutionResult.

ExecutionResult contains exception information and flag value if the
template has failed or not.

Technically this is nice if you want to automate something based on
result.

But in most cases you'd want it to just fail. As it can lead to
behaviour where user is running job  with tempalte. Template is failing
but job succeeds which is going to make user thinking the job is ok. We
have to image that most of the time user haven ot reason to monitor logs
for errors logs.

The user can still catch the exception and handle it. Currently If you
want to handle it in a smart way (e.g ExecutionResult contains the steps
of the template and status of each step. But the exception info does
not. This is the only limitation. It can be fixed in the future by
raising TemplateFailedException which contains template execution result
as a field for users to query. This is not implemented in this change
though.

Testing Done: add functional test that replcates the data job run which
does not fail - after the fix it fails properly.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>