control-service: add enabled flag of deployments to the database (#421)
* control-service: add enabled flag of deployments to the database

Currently, the enabled status of data job deployments is inferred from
their respective cronjob in Kubernetes. This has multiple implications,
including the fact that when we need to know whether a deployment
is enabled, we need to query the Kubernetes API. It is also strange
to not have the complete state of the data jobs in the database and
instead rely for this on Kubernetes.

This commit is a small step towards moving the deployment state entirely
in the database by adding the enabled status of deployments (currently
in the data_job table, but subsequently in a separate table dedicated
to deployments).

Testing done: Manually tested the change by first starting the service
with the new Flyway script and observing that the new column is added
to the data_job table. Validated that the existing jobs have true as
a value of the newly added column. Then, created, and deployed a new job.
Finally disabled, enabled, and deleted the deployment and verified
that the enabled value changed accordingly.

Also, extended the unit tests to verify that the enabled flag is
persisted when it should.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>