control-service: clean up metrics for deleted data jobs
Currently, when a data job is deleted from one instance of the control service,
all other instances continue to expose the metrics for the deleted job.

This commit introduces logic to ensure that metrics are cleaned
up when a job is deleted. This process happens asynchronously on regular
intervals, that are configurable.

Testing done: unit and integration tests. New unit tests added.
Tested the new funcionality in a locally running control service
by removing a row from the database manually and verifying that
the respective metrics get deleted eventually.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>