vdk-core: show default values in vdk config-help
We were defaulting to not showing default values. But that seems
mistake. Default values enable users to more quickly understand the
configuration (as they are an example) and also to see how the vdk would
behave if left unset.

Testing Done: vdk config-help locally and saw they are populated. unit
tests

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>