control-service: Properly classify requirements.txt errors
When a data job is deployed with a misconfigured requirements.txt file (for
example, if a requirement has unexisting version) the job builder pod in k8s
fails, and the error is misclassified as a Platform Error, when it should be
an User Error.

This change ensures that when a ">requirements_failed<" string is present in the
logs, the error is properly classified as an User Error, and the correct exception
message is caught.

Testing Done: Unit test added, and manually tested by deploying a data
job with broken requirements.txt file and verified that it was properly
classified as a User Error.

Signed-off-by: Andon Andonov <andonova@vmware.com>