control-service: add tracking for and automatic release of locks
Currently, if the Control Service is shutdown all locks remain active.
This causes subsequent instances of the service to fail when trying to
obtain said locks and as a result, the functionality behind these locks
is delayed until the previous locks expire.

This change attempts to release all active locks on service' shutdown
by keeping track of all acquired and released locks and handling the
@PreDestroy hook where all active locks are manually released.

Testing done: manually tested the change by running the service,
waiting for a lock to be acquired, then shutting down the service
and verifying that the lock was released (inspecting the database directly).
Also, running the service again has no issues acquiring the lock.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>