control-service: classify DeadlineExceeded status as User Error
Currently, if a data job is running for more than a preconfigured amount (36h at
this time), it is terminated by Kubernetes and the Kubernetes job is marked as failed.
This is interpreted as Platform Error because the pod did not produce a termination
message to classify it otherwise. However, the reason for such a delay is more
often than not an unacceptably slow running data jobs.

This commit classifies executions that ended up exceeding the preconfigured
deadline as User Errors.

Testing done: new unit tests.

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>