vdk-core: define clearly public interfaces (#443)
Document the definiton of public interface.

We slightly difffer from PEP 8 in this
(https://www.python.org/dev/peps/pep-0008/#public-and-internal-interfaces)
so we need to explicitly document it. As there could be internal
documented classes (and we do not explicitly note they are internal).
It's an option to start. But as we have not until now, it should be
separate effort.

Also made some recommendation on hook spec method names.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>