control-service: Update CONTRIBUTING.md
Setting up the control-service environment should be concise,
yet detailed enough, so that the user would not struggle.

Fixed step for importing the project, since the setting.gradle
file is not in the root directory and needs to be linked
separately.

Tested by importing the project and linking it to gradle.

Signed-off-by: Yana Zhivkova <yzhivkova@vmware.com>