vdk-sqlite: Fix sqlite-query bug caused by no cursor description
Sometimes, the result of a SQL query has no description for the
cursor, for example when executing a `DROP TABLE` query. This
causes a `NoneType isn't iterable` error, since we still try
to iterate over the description to get the column names. Note
that this issue was merely a cosmetic one, as the query would
still be executed successfully.
This change fixes this by nesting the printing statements inside
an if-block which checks if the cursor description exists.

Testing done: ran DROP TABLE and observed no errors were caused

Signed-off-by: gageorgiev <gageorgiev@vmware.com>