vdk-sqlite: Fix sqlite-query bug caused by no cursor description (#453)
Sometimes, the result of a SQL query has no description for the
cursor, for example when executing a `DROP TABLE` query. This
causes a `NoneType isn't iterable` error, since we still try
to iterate over the description to get the column names. Note
that this issue was merely a cosmetic one, as the query would
still be executed successfully.
This change fixes this by setting column_names to `()` - same
as the default value of the headers parameter in the tabulate
function - in the case where `cursor.description` doesn't exist.

Testing done: ran DROP TABLE and observed no errors were caused

Signed-off-by: gageorgiev <gageorgiev@vmware.com>