control-service: fix graphql NPE when sorting null description or team
why: Users reported that when trying to sort by a property that
has a null value in the database, control service returns a 400
bad Request error. After some investigation I determined that
this was caused by an incorrect implementation of the custom
comparators for Description and Team which would throw NPE
when comparing a property with a null value.

what: Changed implementation so that it is null tolerant. Fixed indentation
and removed unused imports in the classes that needed changing.

testing: Added two new unit tests. Tested on a locally running control-service
with the same request that previously returned 400 and it seems to
be working as expected after the changes.

Signed-off-by: Momchil Zhivkov mzhivkov@vmware.com