control-service: Remove logging of credentials in builder
Currently, we are leaking AWS credentials in builder pods. These
are then logged, which is a security concern.

This change removes the leaked credentials from the logs.

Testing Done: Manually changed the version of the builder job to a
dev version, in order to push an image to docker hub. Afterwards, started
a local instance of Control Service, and manually deployed a test data job
and verified that the AWS credentials are no longer being logged.

Signed-off-by: Andon Andonov <andonova@vmware.com>