control-service: return vdkVersion in Deployment
VDK version is the version of the VDK library with which the data job
run. We need it to return it when querying for deployment informatin in
order to know with which version of vdk the data job is configured to
run in cloud(kubernetes) deployment.

GET deployment (or list of deployments) returns the configured vdk
version

Testing Done: integration tests, unit tests.

Signed-off-by: Antoni Ivanov <aivanov@vmware.com>