contro-service: revert some of the logic related to false-positive emails
Last batch of changes to handle false-positive notifications introduced
logic that skips updating the last_job_termination_status of a data job
if it is NONE or null.

This was incorrect because of how we find the execution_id inside rules: by
cross-referencing to the last kube_job_status_start_time metric. In this case, if
the last execution terminated with an error and a new execution starts, we will
cross-reference the last termination status (error) with the new job execution
and will send a notification.

This commit reverts the old logic that sents the last_job_termination_status
to NONE when a new execution starts.

Testing done: unit tests

Signed-off-by: Tsvetomir Palashki <tpalashki@vmware.com>